Simplify poll results toggle
